// src/api/generated/models/permission.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gia tot example
 * The gia tot API description
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Permission
 */
export interface Permission {
  /**
   * ID permission
   * @type {string}
   * @memberof Permission
   */
  id: string;
  /**
   * Module
   * @type {string}
   * @memberof Permission
   */
  module: string;
  /**
   * Hành động CRUD
   * @type {string}
   * @memberof Permission
   */
  action: PermissionActionEnum;
  /**
   *
   * @type {string}
   * @memberof Permission
   */
  description?: string;
  /**
   * Permission hệ thống (readonly)
   * @type {boolean}
   * @memberof Permission
   */
  isSystem: boolean;
  /**
   *
   * @type {string}
   * @memberof Permission
   */
  createdAt: string;
  /**
   *
   * @type {string}
   * @memberof Permission
   */
  updatedAt: string;
}

export const PermissionActionEnum = {
  Create: "create",
  Read: "read",
  Update: "update",
  Delete: "delete",
  Cancel: "cancel",
  Publish: "publish",
  AssignRole: "assign.role",
} as const;

export type PermissionActionEnum =
  (typeof PermissionActionEnum)[keyof typeof PermissionActionEnum];